cleaned up and moved logic to correct places
