+++ src/containers/TokenPicker/index.ts
@@ -1,13 +1,12 @@
 import { connect } from 'react-redux'
 import TokenPicker, { TokenPickerProps } from 'components/TokenPicker'
-import { continueToOrder } from 'actions'
-import { setTokenListType } from 'containers/TokenUpload'
+import { continueToOrder, setTokenListType } from 'actions'
 
 import { State } from 'types'
 
 const mapState = ({ ipfs: { fileBuffer, fileHash }, tokenList: { customTokenList, type } }: State) => ({
   fileBuffer,
-  needsTokens: type === 'UPLOAD' || !(fileHash && customTokenList),
+  needsTokens: type === 'UPLOAD' || !(fileHash && customTokenList.length > 0),
 })
 
 export default connect<Partial<TokenPickerProps>, {}, Pick<TokenPickerProps, 'to'>>
